Address review feedback on application_user
